refactor MergeabilityRules.scala to be reusable and put mergeability in separate ci job
